--- src/main/java/com/team25/backend/jwt/JWTFilter.java
@@ -1,82 +1,0 @@
-package com.team25.backend.jwt;
-
-import com.team25.backend.dto.CustomUserDetails;
-import com.team25.backend.entity.User;
-import com.team25.backend.exception.CustomException;
-import com.team25.backend.exception.ErrorCode;
-import com.team25.backend.repository.UserRepository;
-import io.jsonwebtoken.ExpiredJwtException;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-@Slf4j
-public class JWTFilter extends OncePerRequestFilter {
-
-    private final JWTUtil jwtUtil;
-    private final UserRepository userRepository;
-
-    public JWTFilter(JWTUtil jwtUtil, UserRepository userRepository) {
-        this.jwtUtil = jwtUtil;
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        String accessToken = request.getHeader("Authorization");
-
-        if (checkHeader(accessToken)) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        accessToken = accessToken.split(" ")[1];
-
-        try {
-            jwtUtil.isExpired(accessToken);
-        } catch (ExpiredJwtException e) {
-            PrintWriter writer = response.getWriter();
-            writer.print("access token expired");
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
-        }
-
-        String category = jwtUtil.getCategory(accessToken);
-
-        if (!category.equals("access")) {
-            PrintWriter writer = response.getWriter();
-            writer.print("invalid access token");
-
-            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            return;
-        }
-
-        String uuid = jwtUtil.getUuid(accessToken);
-        User foundUser = userRepository.findByUuid(uuid)
-                .orElseThrow(()-> new CustomException(ErrorCode.USER_NOT_FOUND));
-
-        String username = foundUser.getUsername();
-        String role = foundUser.getRole();
-
-        User user = new User(username, uuid, role);
-        CustomUserDetails customUserDetails = new CustomUserDetails(user);
-
-        Authentication authToken = new UsernamePasswordAuthenticationToken(customUserDetails, null, customUserDetails.getAuthorities());
-        SecurityContextHolder.getContext().setAuthentication(authToken);
-
-        filterChain.doFilter(request, response);
-    }
-
-    private boolean checkHeader(String authorization) {
-        return authorization == null || !authorization.startsWith("Bearer ");
-    }
-}
