+++ src/main/java/com/team25/backend/service/JWTService.java
@@ -3,7 +3,7 @@ package com.team25.backend.service;
 import com.team25.backend.dto.response.TokenResponse;
 import com.team25.backend.dto.response.UserResponse;
 import com.team25.backend.entity.Refresh;
-import com.team25.backend.jwt.JWTUtil;
+import com.team25.backend.security.JWTUtil;
 import com.team25.backend.repository.RefreshRepository;
 import org.springframework.stereotype.Service;
